+++ src/components/files/FileToolbar/FileToolbar.tsx
@@ -1,22 +1,86 @@
-import { FC } from "react";
-import { downloadCurrentFile } from "models/levelsets";
-import { Button } from "ui/button";
+import { useStore } from "effector-react";
+import { FC, useCallback } from "react";
+import {
+  $currentFileHasLocalOptions,
+  $hasOtherFiles,
+  downloadCurrentFile,
+} from "models/levelsets";
+import { Button, ButtonDropdown, Toolbar } from "ui/button";
 import { ColorType } from "ui/types";
-import { svgs } from "ui/icon";
+import { IconStack, IconStackType, svgs } from "ui/icon";
 import { useFileButtonsProps } from "./useFileButtonsProps";
 
-export const FileToolbar: FC = () => {
+const removeOtherStack: IconStack = [[IconStackType.Index, <svgs.Trash />]];
+
+interface Props {
+  isCompact?: boolean;
+}
+
+export const FileToolbar: FC<Props> = ({ isCompact = false }) => {
+  const hasLocalOptions = useStore($currentFileHasLocalOptions);
+  const hasOtherFiles = useStore($hasOtherFiles);
   const { filename, handlers } = useFileButtonsProps();
+
+  const saveButton = (
+    <Button
+      uiColor={ColorType.SUCCESS}
+      icon={<svgs.Save />}
+      disabled={!filename}
+      title={filename ? `Save file "${filename}" from memory` : undefined}
+      onClick={useCallback(() => downloadCurrentFile(), [])}
+    />
+  );
+  const saveWithOptionsTitle = filename
+    ? `Save file "${filename}.zip" with Options`
+    : undefined;
+  const handleSaveWithOptions = useCallback(
+    () => downloadCurrentFile({ withLocalOptions: true }),
+    [],
+  );
+
+  const removeButton = (
+    <Button
+      uiColor={ColorType.DANGER}
+      icon={<svgs.Trash />}
+      disabled={!filename}
+      title={filename ? `Remove levelset "${filename}" from memory` : undefined}
+      onClick={handlers?.remove}
+    />
+  );
+  const removeOthersTitle = `Remove others but "${filename}"`;
 
   return (
     <>
-      <Button
-        uiColor={ColorType.SUCCESS}
-        icon={<svgs.Save />}
-        disabled={!filename}
-        title={filename ? `Save file "${filename}" from memory` : undefined}
-        onClick={downloadCurrentFile}
-      />
+      {hasLocalOptions && !isCompact && filename ? (
+        <ButtonDropdown
+          standalone={saveButton}
+          buttonProps={{ uiColor: ColorType.SUCCESS }}
+        >
+          <Toolbar>
+            <Button
+              uiColor={ColorType.SUCCESS}
+              icon={<svgs.Save />}
+              onClick={handleSaveWithOptions}
+            >
+              {saveWithOptionsTitle}
+            </Button>
+          </Toolbar>
+        </ButtonDropdown>
+      ) : (
+        <>
+          {saveButton}
+          {hasLocalOptions && (
+            <Button
+              uiColor={ColorType.SUCCESS}
+              icon={<svgs.Save />}
+              disabled={!filename}
+              title={saveWithOptionsTitle}
+              onClick={handleSaveWithOptions}
+            />
+          )}
+        </>
+      )}
+
       <Button
         uiColor={ColorType.SUCCESS}
         icon={<svgs.FileConvert />}
@@ -30,15 +94,37 @@ export const FileToolbar: FC = () => {
         title="Rename file"
         onClick={handlers?.rename}
       />
-      <Button
-        uiColor={ColorType.DANGER}
-        icon={<svgs.Trash />}
-        disabled={!filename}
-        title={
-          filename ? `Remove levelset "${filename}" from memory` : undefined
-        }
-        onClick={handlers?.remove}
-      />
+
+      {hasOtherFiles && !isCompact ? (
+        <ButtonDropdown
+          standalone={removeButton}
+          buttonProps={{ uiColor: ColorType.DANGER }}
+        >
+          <Toolbar>
+            <Button
+              uiColor={ColorType.DANGER}
+              icon={<svgs.Trash />}
+              iconStack={removeOtherStack}
+              onClick={handlers?.removeOthers}
+            >
+              {removeOthersTitle}
+            </Button>
+          </Toolbar>
+        </ButtonDropdown>
+      ) : (
+        <>
+          {removeButton}
+          {hasOtherFiles && (
+            <Button
+              uiColor={ColorType.DANGER}
+              icon={<svgs.Trash />}
+              iconStack={removeOtherStack}
+              title={removeOthersTitle}
+              onClick={handlers?.removeOthers}
+            />
+          )}
+        </>
+      )}
     </>
   );
 };